% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simusim.multimodels.R
\name{simusim.bivar}
\alias{simusim.bivar}
\title{Simulate simultaneous power for multiple tests in separate models}
\usage{
simusim.multimodels(n = NULL, es = NULL, es1 = NULL, es2 = NULL)
}
\arguments{
\item{n}{Set the size of each sample to be drawn from the population. This is the sample size for which you are estimating statistical power. In other words, setting n to equal 100 will estimate statistical power at n = 100. Accepts any positive whole number smaller than your population.}

\item{es}{Set the units in which you are specifying your effect sizes. Accepts "d" for Cohen's d, "r" for correlation coefficients, and "r2" for percent of variance accounted for.}

\item{models}{How many models would you like to simulate and ultimately calculate power for? Default = 2. Accepts whole numbers in the range of 2 to 10. Note that this argument is required if you specify more than 2 effect sizes.}

\item{null_effect}{For which, if any, of your models are you computing "null power?" If you want to compute "power" to NOT detect any effects, use this argument to specify them by setting this argument equal to the number(s) corresponding to the models you hypothesize to be null. Accepts either a single whole number between 1 and the number of models you have specified or a vector of numbers between 1 and the the number of models you have specified.}

\item{popsize}{What is the size of the population you would like to simulate? This is the population from which you will ultimately draw your samples. Default = 100,000. Accepts any positive whole number}

\item{iterations}{How many times you would like to estimate your models in random samples drawn from your population? One model will be run in each random sample. Default = 5,000. Accepts any whole number greater than 0.}

\item{alpha}{Set your alpha level. This is the threshold below which p-values will be considered significant. Default = 0.05. Accepts any number greater than 0 and less than 1.}

\item{seed}{Set a seed to make your results reproducible. Default = 1. Accepts any number.}

\item{es1...es10}{The effect size, expressed in units specified in the es argument, for each model. You should specify a number of "es"s equal to the number you specified in the "models" argument. That is, if you set models to equal 4, you should supply values for es1, es2, es3, and es4. You should always specify these in order, beginning with es1, and not skipping any. Accepts any number.}
}
\value{
A dataframe containing a power estiamte, expressed as a decimal, for each of the models individually, and for all the models simultaneously.
}
\description{
Simulate power to simultaneously detect significant effects for a set of statistical tests in separate bivariate models. This function simulates data based on a correlation matrix imposed using the corrvar function from the SimCorrMix package (Fialkowski, 2018) using Fleishman's third-order polynomial transformation (Fleishman, 1978), and can be used to estimate power for up to 10 bivariate models with single independent variables and single dependent variables.
}
\details{
When you use this function (and we hope you do!), please cite the package:

Le Forestier, J. M. (2020). SimuSim: Simultaneous power analysis for a set of statistical tests \link{Computer Software}.

and/or cite the accompanying paper:

Le Forestier, J. M., Page-Gould, E., & Chasteen, A. L. (Forthcoming). Statistical power for a set of tests.
}
\examples{
# A basic example, leaving all the defaults in place.

simusim.bivar(n = 150, es = "r", es1 = .2, es2 = .45)

# Another example, customizing additional parameters.

simusim.bivar(n = 300, es = "r2", es1 = .04, es2 = .01, es3 = .00, models = 3,
     null_effect = 3, popsize = 500000, alpha = .01, seed = 123)

}
\references{
Fialkowski, A. C. (2018). SimmCorrMix: Simulation of correlated data with multiple variable types including continuous and count mixture distributions. Comprehensive R Archive Network (CRAN).

Fleishman, A. I. (1978). A method for simulating non-normal distributions. Psychometrika, 43, 521-532.

Le Forestier, J. M., Page-Gould, E., & Chasteen, A. (Forthcoming). Statistical power for a set of tests.
}
\author{
Joel Le Forestier (joel.leforestier@mail.utoronto.ca)
}
