% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simusim.R
\name{simusim}
\alias{simusim}
\title{Simulate simultaneous power for multiple tests}
\usage{
simusim(n = NULL, b1 = NULL, b2 = NULL)
}
\arguments{
\item{predictors}{How many predictor variables would you like to simulate and ultimately calculate power for? Default = 2. Accepts integers in the range of 2 to 10.}

\item{popsize}{What is the size of the population you would like to simulate? This is the population from which you will ultimately draw you samples. Default = 100,000. Accepts any positive integer.}

\item{iterations}{How many times you would like to estimate your models in random samples drawn from your population? One model will be run in each random sample. Default = 10,000. Accepts any positive integer.}

\item{alpha}{Set your alpha level. This is the threshold below which p-values will be considered significant. Default = 0.05. Accepts any number.}

\item{seed}{Set a seed to make your results reproducible. Default = 1. Accepts any number.}

\item{n}{Set the size of each sample to be drawn from the population. This is the sample size for which you are estimating statistical power. In other words, setting n to equal 100 will estimate statistical power at n = 100.}

\item{b1...b10}{The effect size, expressed as correlation coefficients (r), for each predictor in your model. You should specify a number of "b"s equal to the number you specified in the "predictors" argument. That is, if you set predictors to equal 4, you should supple values for b1, b2, b3, and b4. You should always specify these in order, beginning with b1, and not skipping any. You must specify effect sizes for at least b1 and b2, neither of which has a default value. Default values for all other bs are 0. Leave these defaults in place for any unused bs. Accepts any number between -1 and 1.}

\item{iv1iv2_cov...iv9iv10_cov}{The covariance, expressed as correlation coefficients (r), between each set of predictors. Specifying covariances between predictors is optional unless your predictors, together, account for more than 100\% of the variance in your DV, in which case you must specify covariances between your predictors to make that possible. Default = 0. Accepts any number between -1 and 1.}
}
\description{
Simulate power to simultaneously detect significant effects for a set of staitsitcal tests. This function simulates data based on a correlation matrix imposed using the SimMultiCorrData package (Fialkowski, A. C., 2018) using Fleishman's third-order polynomial transformation (Fleishman, 1978).
}
\examples{
# A basic example, leaving all the defaults in place.

simusim(n = 150, b1 = .2, b2 = .15)

# Another example, customizing additional parameters.

simusim(n = 300, b1 = .7, b2 = .6, b3 = .5, iv1iv2_cov = .5,
     popsize = 1000000, alpha = .01, seed = 12345)

}
\references{
Le Forestier, J. M., Page-Gould, E., & Chasteen, A. (Forthcoming). Statistical power for a set of tests.
}
