% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simusim.bivar.R
\name{simusim.bivar}
\alias{simusim.bivar}
\title{Simulate simultaneous power for multiple tests in separate models}
\usage{
simusim.bivar(n = NULL, b1 = NULL, b2 = NULL)
}
\arguments{
\item{n}{Set the size of each sample to be drawn from the population. This is the sample size for which you are estimating statistical power. In other words, setting n to equal 100 will estimate statistical power at n = 100. Accepts any positive whole number smaller than your population.}

\item{models}{How many models would you like to simulate and ultimately calculate power for? Default = 2. Accepts whole numbers in the range of 2 to 10. Note that this argument is required if you specify more than 2 effect sizes.}

\item{popsize}{What is the size of the population you would like to simulate? This is the population from which you will ultimately draw your samples. Default = 100,000. Accepts any positive whole number}

\item{iterations}{How many times you would like to estimate your models in random samples drawn from your population? One model will be run in each random sample. Default = 10,000. Accepts any whole number greater than 0.}

\item{alpha}{Set your alpha level. This is the threshold below which p-values will be considered significant. Default = 0.05. Accepts any number.}

\item{seed}{Set a seed to make your results reproducible. Default = 1. Accepts any number.}

\item{d1...d10}{The effect size, expressed as Cohen's d, for each model. You should specify a number of effect sizes equal to the number you specified in the "models" argument. That is, if you set models to equal 4, you should supply values for d1, d2, d3, and d4. You should always specify these in order, beginning with d1, and not skipping any. You must specify effect sizes for at least d1 and d2. Accepts any number.}
}
\description{
Simulate power to simultaneously detect significant effects for a set of statistical tests in separate bivariate models. This function simulates data based on a correlation matrix imposed using the corrvar function from the SimCorrMix package (Fialkowski, 2018) using Fleishman's third-order polynomial transformation (Fleishman, 1978), and can be used to estimate power for up to 10 bivariate models with single independent variables and single dependent variables.
}
\details{
When you use this function (and we hope you do!), please cite it as:

Le Forestier, J. M., Page-Gould, E., & Chasteen, A. L. (Forthcoming). Statistical power for a set of tests.
}
\examples{
# A basic example, leaving all the defaults in place.

simusim.bivar(n = 150, d1 = .2, d2 = .45)

# Another example, customizing additional parameters.

simusim.bivar(n = 300, d1 = .6, d2 = .43, d3 = .21, models = 3,
     popsize = 500000, alpha = .01, seed = 123)

}
\references{
Fialkowski, A. C. (2018). SimmCorrMix: Simulation of correlated data with multiple variable types including continuous and count mixture distributions. Comprehensive R Archive Network (CRAN).

Fleishman, A. I. (1978). A method for simulating non-normal distributions. Psychometrika, 43, 521-532.

Le Forestier, J. M., Page-Gould, E., & Chasteen, A. (Forthcoming). Statistical power for a set of tests.
}
\author{
Joel Le Forestier (joel.leforestier@mail.utoronto.ca)
}
